compositor|refactor: Consolidate duplicated helpers and conform naming
A helper shared by more than one module now sits at the tier its consumers share, and a helper that had been copied into two modules has a single definition.

`readFileIfPresent` moves to `portable/`, beside the `isMissingFile` it wraps, and `buildConfig` moves from `config/test-utils/` to `src/test-utils/`, so selection's tests no longer reach into config's private helpers. Catalog indexing reuses `portable/appendTo` rather than re-deriving it, and the `tierAt` test helper becomes `buildTierFile`.

Comment lines and one error message now fit the 120-column limit `.editorconfig` sets.
