+++ packages/compositor/src/config/loadConfig.ts
@@ -1,10 +1,9 @@
-import { readFile } from 'node:fs/promises';
 import path from 'node:path';
 
 import { parse as parseYaml } from 'yaml';
 import { z } from 'zod';
 
-import { isMissingFile } from '../portable/isMissingFile.ts';
+import { readFileIfPresent } from '../portable/readFileIfPresent.ts';
 import { IdSchema } from '../schemas/common.ts';
 import type { CompositorConfig, ConfigTier, TierBody } from '../schemas/config-schemas.ts';
 import { TierBodySchema } from '../schemas/config-schemas.ts';
@@ -21,16 +20,16 @@ export type TierFile = z.infer<typeof TierFileSchema>;
 /**
  * Reads the config the given tier files declare, lowest precedence first.
  *
- * Tier identity is the consumer's to supply: a file does not know which tier it is, that follows from where the consumer
- * looked for it, and keeping the decision outside means no particular project layout is compiled in here.
+ * Tier identity is the consumer's to supply: a file does not know which tier it is, that follows from where the
+ * consumer looked for it, and keeping the decision outside means no particular project layout is compiled in here.
  *
  * A tier whose file is absent contributes no tier at all, while one whose file is present but empty contributes a tier
  * declaring nothing. That distinction is what lets a consumer tell "no config here" from "config here, saying nothing".
- * A chain where every file is absent yields a config with no tiers, which already expresses "nothing declared" without a
- * second sentinel.
+ * A chain where every file is absent yields a config with no tiers, which already expresses "nothing declared" without
+ * a second sentinel.
  *
- * Loading is one way to obtain a config, not the only one: the value this returns is the same shape a consumer can build
- * in memory and hand straight to the flows downstream.
+ * Loading is one way to obtain a config, not the only one: the value this returns is the same shape a consumer can
+ * build in memory and hand straight to the flows downstream.
  */
 export async function loadConfig(tiers: ReadonlyArray<TierFile>): Promise<CompositorConfig> {
   const validated = tiers.map((tier) => TierFileSchema.parse(tier));
@@ -64,23 +63,11 @@ function parseTierBody(raw: string, filePath: string): TierBody {
   return result.data;
 }
 
-/** Reads `filePath`, resolving to nothing when it is absent. */
-async function readFileIfPresent(filePath: string): Promise<string | undefined> {
-  try {
-    return await readFile(filePath, 'utf8');
-  } catch (error: unknown) {
-    if (isMissingFile(error)) {
-      return undefined;
-    }
-    throw error;
-  }
-}
-
 /**
  * Reads one tier from its file, resolving to nothing when that file is absent.
  *
- * `baseDir` is the file's own directory, so a relative path a tier declares resolves against where it was written rather
- * than against whatever directory the process happens to be run from.
+ * `baseDir` is the file's own directory, so a relative path a tier declares resolves against where it was written
+ * rather than against whatever directory the process happens to be run from.
  */
 async function readTier(tier: TierFile): Promise<ConfigTier | undefined> {
   const raw = await readFileIfPresent(tier.path);
