+++ packages/compositor/src/test-utils/buildConfig.ts
@@ -1,5 +1,5 @@
-import type { CompositorConfig } from '../../schemas/config-schemas.ts';
-import { CompositorConfigSchema } from '../../schemas/config-schemas.ts';
+import type { CompositorConfig } from '../schemas/config-schemas.ts';
+import { CompositorConfigSchema } from '../schemas/config-schemas.ts';
 
 /**
  * One tier as a test declares it: an authored body, plus an identity defaulted from its position.
